create import component
