+++ packages/client/hmi-client/src/components/documents/Document.vue
@@ -73,12 +73,7 @@
 				<ul>
 					<template v-if="isEditable">
 						<li class="github-link" v-for="(url, index) in githubUrls" :key="index">
-							<Button
-								label="Import"
-								class="p-button-sm p-button-outlined"
-								icon="pi pi-cloud-download"
-								@click="openCodeBrowser(url)"
-							/>
+							<import-code-button :url="url" @open-asset="openAsset" />
 							<a :href="url.html_url" rel="noreferrer noopener">{{ url.html_url }}</a>
 						</li>
 					</template>
@@ -129,22 +124,6 @@
 				</DataTable>
 			</AccordionTab>
 		</Accordion>
-		<Teleport to="body">
-			<modal v-if="isModalVisible" class="modal" @modal-mask-clicked="isModalVisible = false">
-				<template #header>
-					<h5>Choose file to open from {{ chosenRepositoryName }}</h5>
-				</template>
-				<template #default>
-					<ul class="repository-content">
-						<li v-for="(content, index) in filesToSelect" :key="index" @click="openCode(content)">
-							<i v-if="content.download_url === null" class="pi pi-folder" />
-							<i v-else class="pi pi-file" />
-							{{ content.name }}
-						</li>
-					</ul>
-				</template>
-			</modal>
-		</Teleport>
 	</section>
 </template>
 
@@ -160,15 +139,15 @@ import { getDocumentById, getXDDArtifacts } from '@/services/data';
 import { XDDExtractionType } from '@/types/XDD';
 import { XDDArtifact, DocumentType } from '@/types/Document';
 import { getDocumentDoi, isModel, isDataset, isDocument } from '@/utils/data-util';
-import { ResultType } from '@/types/common';
+import { ResultType, Tab } from '@/types/common';
 import { getRelatedArtifacts } from '@/services/provenance';
 import TeraShowMoreText from '@/components/widgets/tera-show-more-text.vue';
+import ImportCodeButton from '@/components/widgets/import-code-button.vue';
 import { Model } from '@/types/Model';
 import { Dataset } from '@/types/Dataset';
 import { ProvenanceType } from '@/types/Provenance';
-import { getGithubUrls, getGithubRepositoryContent, getGithubCode } from '@/services/github-import';
-import modal from '@/components/widgets/Modal.vue';
-import { ProjectAssetTypes } from '@/types/Project';
+import { getGithubUrls } from '@/services/github-import';
+
 import * as textUtil from '@/utils/text';
 
 const props = defineProps<{
@@ -178,15 +157,20 @@ const props = defineProps<{
 	previewLineLimit?: number;
 }>();
 
-const emit = defineEmits(['open-asset']);
-
 const sectionElem = ref<HTMLElement | null>(null);
 const doc = ref<DocumentType | null>(null);
 
-const githubUrls = ref();
-const isModalVisible = ref(false);
-const chosenRepositoryName = ref('');
-const filesToSelect = ref();
+const githubUrls = ref(); // this is will hold github urls temporarily, later we they will be in the document urls
+
+onMounted(async () => {
+	githubUrls.value = await getGithubUrls();
+});
+
+const emit = defineEmits(['open-asset']);
+
+function openAsset(assetToOpen: Tab, newCode?: string) {
+	emit('open-asset', assetToOpen, newCode);
+}
 
 // Highlight strings based on props.highlight
 function highlightSearchTerms(text: string | undefined): string {
@@ -194,31 +178,6 @@ function highlightSearchTerms(text: string | undefined): string {
 		return textUtil.highlight(text, props.highlight);
 	}
 	return text ?? '';
-}
-
-onMounted(async () => {
-	githubUrls.value = await getGithubUrls();
-});
-
-async function openCodeBrowser(url) {
-	isModalVisible.value = true;
-	chosenRepositoryName.value = url.full_name;
-	filesToSelect.value = await getGithubRepositoryContent(url.contents_url.slice(0, -8));
-}
-
-async function openCode(url) {
-	if (url.download_url === null) {
-		chosenRepositoryName.value = `${chosenRepositoryName.value}/${url.name}`;
-		filesToSelect.value = await getGithubRepositoryContent(url.url);
-		return;
-	}
-
-	emit(
-		'open-asset',
-		{ assetName: 'New file', assetType: ProjectAssetTypes.CODE },
-		// { assetName: url.name, assetId: url.name, assetType: ProjectAssetTypes.CODE },
-		await getGithubCode(url.download_url)
-	);
 }
 
 watch(
@@ -353,23 +312,3 @@ onMounted(async () => {
 	fetchRelatedTerariumArtifacts();
 });
 </script>
-
-<style scoped>
-.repository-content {
-	list-style: none;
-	display: flex;
-	flex-direction: column;
-	gap: 0.25rem;
-	margin-top: 1rem;
-}
-
-.repository-content li {
-	padding: 0.25rem;
-	cursor: pointer;
-	border-radius: 0.5rem;
-}
-
-.repository-content li:hover {
-	background-color: var(--surface-hover);
-}
-</style>
